ResolverImpl now uses streams
