+++ src/main/java/com/ddoerr/scriptit/ScriptItMod.java
@@ -5,8 +5,8 @@ import com.ddoerr.scriptit.api.libraries.NamespaceRegistry;
 import com.ddoerr.scriptit.api.bus.EventBus;
 import com.ddoerr.scriptit.api.bus.KeyBindingBusExtension;
 import com.ddoerr.scriptit.config.ConfigHandler;
-import com.ddoerr.scriptit.dependencies.Loadable;
-import com.ddoerr.scriptit.dependencies.Resolver;
+import com.ddoerr.scriptit.api.dependencies.Loadable;
+import com.ddoerr.scriptit.api.dependencies.Resolver;
 import com.ddoerr.scriptit.elements.HudElementManager;
 import com.ddoerr.scriptit.loader.EventLoader;
 import com.ddoerr.scriptit.loader.HudElementLoader;
