+++ src/main/java/com/ddoerr/scriptit/api/bus/KeyBindingBusExtension.java
@@ -2,7 +2,7 @@ package com.ddoerr.scriptit.api.bus;
 
 import com.ddoerr.scriptit.ScriptItMod;
 import com.ddoerr.scriptit.api.util.KeyBindingHelper;
-import com.ddoerr.scriptit.dependencies.Resolver;
+import com.ddoerr.scriptit.api.dependencies.Resolver;
 import net.minecraft.client.options.KeyBinding;
 import net.minecraft.client.util.InputUtil;
 import net.minecraft.util.Identifier;
