+++ src/main/java/com/ddoerr/scriptit/api/dependencies/Loadable.java
@@ -1,4 +1,4 @@
-package com.ddoerr.scriptit.dependencies;
+package com.ddoerr.scriptit.api.dependencies;
 
 public interface Loadable {
     void load();
