+++ src/main/java/com/ddoerr/scriptit/config/HudElementAdapter.java
@@ -5,7 +5,7 @@ import com.ddoerr.scriptit.elements.HudElement;
 import com.ddoerr.scriptit.scripts.ScriptContainer;
 import com.google.gson.*;
 import com.ddoerr.scriptit.api.util.geometry.Point;
-import com.ddoerr.scriptit.dependencies.Resolver;
+import com.ddoerr.scriptit.api.dependencies.Resolver;
 import com.ddoerr.scriptit.loader.HudElementLoader;
 
 import java.lang.reflect.Type;
