+++ src/main/java/com/ddoerr/scriptit/libraries/ScriptsLibrary.java
@@ -3,7 +3,7 @@ package com.ddoerr.scriptit.libraries;
 import com.ddoerr.scriptit.api.libraries.LibraryInitializer;
 import com.ddoerr.scriptit.api.libraries.LibraryRegistry;
 import com.ddoerr.scriptit.api.libraries.NamespaceRegistry;
-import com.ddoerr.scriptit.dependencies.Resolver;
+import com.ddoerr.scriptit.api.dependencies.Resolver;
 import com.ddoerr.scriptit.scripts.ThreadLifetimeManager;
 
 public class ScriptsLibrary implements LibraryInitializer {
