+++ src/main/java/com/ddoerr/scriptit/loader/LanguageLoader.java
@@ -2,7 +2,7 @@ package com.ddoerr.scriptit.loader;
 
 import com.ddoerr.scriptit.ScriptItMod;
 import com.ddoerr.scriptit.api.languages.LanguageRegistry;
-import com.ddoerr.scriptit.dependencies.Loadable;
+import com.ddoerr.scriptit.api.dependencies.Loadable;
 import com.ddoerr.scriptit.api.languages.LanguageImplementation;
 import com.ddoerr.scriptit.api.languages.LanguageInitializer;
 import net.fabricmc.loader.api.FabricLoader;
