+++ src/main/java/com/ddoerr/scriptit/triggers/BusTrigger.java
@@ -5,7 +5,7 @@ import com.ddoerr.scriptit.api.bus.Bus;
 import com.ddoerr.scriptit.api.bus.EventBus;
 import com.ddoerr.scriptit.api.libraries.NamespaceRegistry;
 import com.ddoerr.scriptit.api.bus.KeyBindingBusExtension;
-import com.ddoerr.scriptit.dependencies.Resolver;
+import com.ddoerr.scriptit.api.dependencies.Resolver;
 import net.minecraft.client.util.InputUtil;
 
 import java.util.function.Consumer;
